users

+++ src/users/users.controller.ts
@@ -3,6 +3,8 @@ import {
   Controller,
   Delete,
   Get,
+  HttpException,
+  HttpStatus,
   Param,
   ParseIntPipe,
   Patch,
@@ -11,33 +13,56 @@ import {
 import { UsersService } from './users.service';
 import { createUserDto } from './dto/create-user.dto';
 import { updateUserDto } from './dto/update-user.dto';
+import { User } from './user.entity';
 
 @Controller('users')
 export class UsersController {
   constructor(private userService: UsersService) {}
 
   @Post()
-  createUser(@Body() data: createUserDto) {
+  async createUser(@Body() data: createUserDto): Promise<User | HttpException> {
+    if (!data.email || !data.password || !data.username)
+      throw new HttpException('missing values', HttpStatus.BAD_REQUEST);
+    const findUser = await this.userService.getUserForCreate({
+      email: data.email,
+      username: data.username,
+    });
+    if (findUser)
+      throw new HttpException('user already exists', HttpStatus.CONFLICT);
     return this.userService.createUser(data);
   }
 
   @Get()
-  getUsers() {
+  getUsers(): Promise<User[]> {
     return this.userService.getUsers();
   }
 
   @Get(':id')
-  getUser(@Param('id', ParseIntPipe) id: number) {
+  async getUser(
+    @Param('id', ParseIntPipe) id: number,
+  ): Promise<User | HttpException> {
+    const findUser = await this.userService.getUser(id);
+    if (!findUser)
+      throw new HttpException('user not found', HttpStatus.NOT_FOUND);
     return this.userService.getUser(id);
   }
 
   @Patch(':id')
-  updateUser(@Param('id') id: number, @Body() data: updateUserDto) {
+  async updateUser(
+    @Param('id') id: number,
+    @Body() data: updateUserDto,
+  ): Promise<User> {
+    const findUser = await this.userService.getUser(id);
+    if (!findUser)
+      throw new HttpException('user not found', HttpStatus.NOT_FOUND);
     return this.userService.updateUser(id, data);
   }
 
   @Delete(':id')
-  deleteUser(@Param('id') id: number) {
+  async deleteUser(@Param('id') id: number) {
+    const findUser = await this.userService.getUser(id);
+    if (!findUser)
+      throw new HttpException('user not found', HttpStatus.NOT_FOUND);
     return this.userService.deleteUser(id);
   }
 }
